fix(ui): use one direct media tab opener
