--- invokeai/frontend/web/src/common/util/openImageInNewTab.ts
@@ -1,3 +1,0 @@
-export const openImageInNewTab = (imageUrl: string) => {
-  window.open(imageUrl, '_blank', 'noopener,noreferrer');
-};
