+++ invokeai/frontend/web/src/common/util/openMediaInNewTab.ts
@@ -1,0 +1,3 @@
+export const openMediaInNewTab = (mediaUrl: string) => {
+  window.open(mediaUrl, '_blank', 'noopener,noreferrer');
+};
