+++ invokeai/frontend/web/src/features/auth/hooks/useMediaCookieRefresh.test.ts
@@ -1,21 +1,6 @@
-import { afterEach, describe, expect, it, vi } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 
-import { abortAndWaitForPendingRefreshes, openMediaInNewTab } from './useMediaCookieRefresh';
-
-describe('openMediaInNewTab', () => {
-  afterEach(() => {
-    vi.unstubAllGlobals();
-  });
-
-  it('opens the media URL directly when cookie self-heal is not pending', () => {
-    const open = vi.fn(() => null);
-    vi.stubGlobal('window', { open });
-
-    openMediaInNewTab('api/v1/images/i/test.png/full');
-
-    expect(open).toHaveBeenCalledWith('api/v1/images/i/test.png/full', '_blank', 'noopener,noreferrer');
-  });
-});
+import { abortAndWaitForPendingRefreshes } from './useMediaCookieRefresh';
 
 describe('media cookie refresh shutdown', () => {
   it('aborts a stalled request instead of blocking logout indefinitely', async () => {
