+++ invokeai/frontend/web/src/features/gallery/components/ContextMenu/MenuItems/ContextMenuItemOpenInNewTab.tsx
@@ -1,5 +1,5 @@
 import { IconMenuItem } from 'common/components/IconMenuItem';
-import { openImageInNewTab } from 'common/util/openImageInNewTab';
+import { openMediaInNewTab } from 'common/util/openMediaInNewTab';
 import { useImageDTOContext } from 'features/gallery/contexts/ImageDTOContext';
 import { memo, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -9,7 +9,7 @@ export const ContextMenuItemOpenInNewTab = memo(() => {
   const { t } = useTranslation();
   const imageDTO = useImageDTOContext();
   const onClick = useCallback(() => {
-    openImageInNewTab(imageDTO.image_url);
+    openMediaInNewTab(imageDTO.image_url);
   }, [imageDTO]);
 
   return (
